updated core event tests
