+++ tests/core/course_module_completion_update/completing_existing_module/completing_existing_module_test.php
@@ -14,10 +14,24 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace logstore_xapi\core\course_viewed\viewing_existing_course;
+namespace logstore_xapi\core\course_module_completion_update\completing_existing_module;
 
-class test extends \logstore_xapi\xapi_test_case {
+defined('MOODLE_INTERNAL') || die();
+
+global $CFG;
+
+require_once($CFG->dirroot . '/admin/tool/log/store/xapi/vendor/autoload.php');
+require_once($CFG->dirroot . '/admin/tool/log/store/xapi/src/autoload.php');
+require_once($CFG->dirroot . '/admin/tool/log/store/xapi/tests/xapi_test_case.php');
+
+class completing_existing_module_test extends \logstore_xapi\xapi_test_case {
     protected function get_test_dir() {
         return __DIR__;
+    }
+
+    // Adding this to appease the auto-detecting of test cases.
+    // xapi_test_case has the default test cases and this test extends that.
+    public function test_init() {
+
     }
 }
