+++ tests/core/user_loggedout/existing_user_loggedout/existing_user_loggedout_test.php
@@ -14,18 +14,16 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace logstore_xapi\core\course_module_completion_update\completing_existing_module;
+namespace logstore_xapi\core\user_loggedout\existing_user_loggedout;
 
-defined('MOODLE_INTERNAL') || die();
-
-global $CFG;
-
-require_once($CFG->dirroot . '/admin/tool/log/store/xapi/vendor/autoload.php');
-require_once($CFG->dirroot . '/admin/tool/log/store/xapi/src/autoload.php');
-require_once($CFG->dirroot . '/admin/tool/log/store/xapi/tests/xapi_test_case.php');
-
-class test extends \logstore_xapi\xapi_test_case {
+class existing_user_loggedout_test extends \logstore_xapi\xapi_test_case {
     protected function get_test_dir() {
         return __DIR__;
+    }
+
+    // Adding this to appease the auto-detecting of test cases.
+    // xapi_test_case has the default test cases and this test extends that.
+    public function test_init() {
+
     }
 }
